requirements.txt
aiogram
asyncpg
python-dotenv
apscheduler
pytz
openpyxl